# apiGateway/src/swagger/swagger.json
{
  "basePath": "",
  "host": "localhost:3000",
  "info": {
    "description": "SFA API",
    "version": "1.0.0",
    "title": "SFA API",
    "termsOfService": "",
    "contact": {
      "email": "support@troondx.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "schemes": [
    "https",
    "http"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "x-access-token",
      "in": "header",
      "description": "<token>"
    }
  },
  "swagger": "2.0",
  "paths": {
    "/auth/adminlogin": {
      "post": {
        "description": "AdminLogin from Admin Web Portal.",
        "operationId": "adminlogin",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object"
            }
          }
        },
        "summary": "Set Response token as Authorize Api-Key ( Value:<token> )",
        "tags": [
          "User"
        ]
      }
    },
...
    "/apia/getUserTypes": {
      "post": {
        "description": "Get List of User Types",
        "operationId": "UserTypes",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/filter"
            }
          }
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "security": [
          {
            "api_key": []
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object"
            }
          }
        },
        "summary": "Fetch List of User Types in the system.",
        "tags": [
          "Lookup"
        ]
      }
    },
    "/apia/getUserRoles": {
      "post": {
        "description": "Get List of User Roles",
        "operationId": "UserRoles",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/filter"
            }
          }
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "security": [
          {
            "api_key": []
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object"
            }
          }
        },
        "summary": "Fetch List of User Roles in the system.",
        "tags": [
          "Lookup"
        ]
      }
    },
...
}
  }